+++ src/endstone_ninjos_schematics/__init__.py
@@ -1,6 +1,6 @@
 """Ninj-OS Schematics plugin package."""
 
-__version__ = "1.7.1"
+__version__ = "1.7.2"
 
 try:
     from .plugin import NinjOSSchematicsPlugin